move funkyPlugin.js to index.js
