+++ package.json
@@ -15,7 +15,7 @@
     "test": "npm run test:unit && npm run test:typescript",
     "test:unit": "jest --config=jest.config.json --coverage",
     "test:typescript": "tsd",
-    "lint": "eslint \"lib/**/*.js\" \"test/**/*.js\" index.js",
+    "lint": "eslint \"test/**/*.js\" index.js",
     "prettier": "prettier --write \"{lib,test}/**/*.js\" index.js index.d.ts"
   },
   "dependencies": {
